+++ org.mmisw.orrportal/src/org/mmisw/orrclient/OrrClientFactory.java
@@ -6,8 +6,8 @@ import org.mmisw.orrclient.core.OrrClientImpl;
  * Obtains the IOrrClient implementation
  */
 public class OrrClientFactory {
-  public static IOrrClient init(OrrClientConfiguration config) throws Exception {
-    return OrrClientImpl.init(config);
+  public static IOrrClient init() throws Exception {
+    return OrrClientImpl.init();
   }
 
   public static IOrrClient getOrrClient() {