+++ org.mmisw.orrportal/src/org/mmisw/orrclient/core/vine/MappingOntologyCreator.java
@@ -1,5 +1,6 @@
 package org.mmisw.orrclient.core.vine;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.List;
@@ -12,13 +13,12 @@ import org.mmisw.ont.JenaUtil2;
 import org.mmisw.ont.vocabulary.Omv;
 import org.mmisw.ont.vocabulary.OmvMmi;
 import org.mmisw.ont.vocabulary.Vine;
-import org.mmisw.orrclient.OrrClientConfiguration;
-import org.mmisw.orrclient.OrrClientFactory;
 import org.mmisw.orrclient.core.MdHelper;
 import org.mmisw.orrclient.gwt.client.rpc.CreateOntologyInfo;
 import org.mmisw.orrclient.gwt.client.rpc.CreateOntologyResult;
 import org.mmisw.orrclient.gwt.client.rpc.MappingDataCreationInfo;
 import org.mmisw.orrclient.gwt.client.rpc.vine.Mapping;
+import org.mmisw.orrportal.gwt.server.OrrConfig;
 
 import com.hp.hpl.jena.ontology.OntModel;
 import com.hp.hpl.jena.ontology.Ontology;
@@ -70,13 +70,6 @@ public class MappingOntologyCreator {
 	private final String uniqueBaseName = _createUniqueBaseName();
 
 
-	private static OrrClientConfiguration _config() {
-		OrrClientConfiguration config = OrrClientFactory.getOrrClient().getConfiguration();
-		return config;
-	}
-
-
-
 	/**
 	 *
 	 * @param createOntologyInfo      Metadata for the ontology
@@ -92,7 +85,7 @@ public class MappingOntologyCreator {
 		this.namespaces = mappingDataCreationInfo.getNamespaces();
 		this.mappings = mappingDataCreationInfo.getMappings();
 
-		String ontServiceUrl = _config().getOntServiceUrl();
+		String ontServiceUrl = OrrConfig.instance().ontServiceUrl;
 		this.namespaceRoot = ontServiceUrl;
 		this.orgAbbreviation = createOntologyInfo.getAuthority();
 		this.shortName = createOntologyInfo.getShortName();
@@ -136,7 +129,7 @@ public class MappingOntologyCreator {
 
 
 		// now save the RDF:
-		String full_path = _config().getVoc2rdfDirectory() + getPathOnServer();
+		File full_path = new File(OrrConfig.instance().voc2rdfDir, getPathOnServer());
 		try {
 			FileWriter os = new FileWriter(full_path);
 			os.write(rdf);
@@ -150,7 +143,7 @@ public class MappingOntologyCreator {
 		}
 
 		// OK:
-		createVocabResult.setFullPath(full_path);
+		createVocabResult.setFullPath(full_path.getAbsolutePath());
 	}
 
 	private OntModel _createOntModel() {
@@ -271,7 +264,6 @@ public class MappingOntologyCreator {
 
 	/**
 	 * Creates and adds to the model the mapping statement.
-	 * @param mapping
 	 * @param r
 	 * @param p
 	 * @param o