+++ org.mmisw.orrportal/src/org/mmisw/orrportal/gwt/client/rpc/PortalBaseInfo.java
@@ -4,70 +4,33 @@ import java.io.Serializable;
 
 /**
  * Base info for the portal.
- * 
- * @author Carlos Rueda
  */
 public class PortalBaseInfo implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
-	private String appServerUrl;
 	private String ontServiceUrl;
-	private String ontbrowserServiceUrl;
-	private String gaUaNumber;
-	
-	
-	/**
-	 * @return the appServerUrl
-	 */
-	public String getAppServerUrl() {
-		return appServerUrl;
-	}
-	/**
-	 * @param appServerUrl the appServerUrl to set
-	 */
-	public void setAppServerUrl(String appServerUrl) {
-		this.appServerUrl = appServerUrl;
-	}
+	private String touUrl;
+
+	public PortalBaseInfo() {}
+
 	/**
 	 * @return the ontServiceUrl
 	 */
 	public String getOntServiceUrl() {
 		return ontServiceUrl;
 	}
-	/**
-	 * @param ontServiceUrl the ontServiceUrl to set
-	 */
-	public void setOntServiceUrl(String ontServiceUrl) {
-		this.ontServiceUrl = ontServiceUrl;
-	}
+  /**
+   * @param ontServiceUrl the ontServiceUrl to set
+   */
+  public void setOntServiceUrl(String ontServiceUrl) {
+    this.ontServiceUrl = ontServiceUrl;
+  }
 
-	/**
-	 * @return the ontbrowserServiceUrl
-	 */
-	public String getOntbrowserServiceUrl() {
-		return ontbrowserServiceUrl;
-	}
-	/**
-	 * @param ontbrowserServiceUrl the ontbrowserServiceUrl to set
-	 */
-	public void setOntbrowserServiceUrl(String ontbrowserServiceUrl) {
-		this.ontbrowserServiceUrl = ontbrowserServiceUrl;
-	}
+  public String getTouUrl() {
+    return touUrl;
+  }
 
-	/**
-	 * Sets the google analytics UA number
-	 * @param gaUaNumber
-	 */
-	public void setGaUaNumber(String gaUaNumber) {
-		this.gaUaNumber = gaUaNumber;
-	}
-	
-	/**
-	 * Gets the google analytics UA number
-	 * @return the google analytics UA number
-	 */
-	public String getGaUaNumber() {
-		return gaUaNumber;
-	}
-	
+  public void setTouUrl(String touUrl) {
+    this.touUrl = touUrl;
+  }
 }